added solver time ranges to SGH

+++ single-node-refactor/src/solver.h
@@ -53,8 +53,8 @@ public:
 
     size_t solver_id;
 
-    size_t time_start;
-    size_t time_end;
+    double time_start;
+    double time_end;
 
     Solver();
     virtual ~Solver();
